Struct ethcore_signer::ServerBuilder

pub struct ServerBuilder { /* fields omitted */ }

Builder for WebSockets server

Methods

impl ServerBuilder

fn new(queue: Arc<ConfirmationsQueue>, authcodes_path: PathBuf) -> Self

Creates new ServerBuilder

fn skip_origin_validation(self, skip: bool) -> Self

If set to true server will not verify Origin of incoming requests. Not recommended. Use only for development.

fn start(self, addr: SocketAddr) -> Result<Server, ServerError>

Starts a new WebSocket server in separate thread. Returns a Server handle which closes the server when droped.